+++ src/starbase/types.ts
@@ -11,16 +11,33 @@ export type StarbaseStorage<TStorageConfig = any> = {
   config: TStorageConfig;
 };
 
-interface Neo4jStorageConfig {
+interface Neo4jStorageEngineConfig {
   username: string;
   password: string;
   uri: string;
   database?: string;
 }
 
-export type Neo4jStorage = StarbaseStorage<Neo4jStorageConfig>;
+interface JupiterOneStorageEngineConfig {
+  /**
+   * The JupiterOne API key to authenticate with
+   */
+  apiKey: string;
+  /**
+   * The JupiterOne account ID to target
+   */
+  accountId: string;
+  /**
+   * If provided this option specifies which base URL to use for synchronization (default: `https://api.us.jupiterone.io`)
+   */
+  apiBaseUrl?: string;
+}
+
+export type Neo4jStorage = StarbaseStorage<Neo4jStorageEngineConfig>;
+
+export type JupiterOneStorage = StarbaseStorage<JupiterOneStorageEngineConfig>;
 
 export type StarbaseConfig = {
   integrations: StarbaseIntegration[];
-  storage?: StarbaseStorage;
+  storage?: StarbaseStorage[];
 };